+++ packages/opentelemetry-sdk-trace-base/src/Tracer.ts
@@ -20,15 +20,14 @@ import {
   sanitizeAttributes,
   isTracingSuppressed,
 } from '@opentelemetry/core';
-import { IResource } from '@opentelemetry/resources';
-import { BasicTracerProvider } from './BasicTracerProvider';
-import { Span } from './Span';
+import { SpanImpl } from './Span';
 import { GeneralLimits, SpanLimits, TracerConfig } from './types';
 import { mergeConfig } from './utility';
 import { SpanProcessor } from './SpanProcessor';
 import { Sampler } from './Sampler';
 import { IdGenerator } from './IdGenerator';
 import { RandomIdGenerator } from './platform';
+import { IResource } from '@opentelemetry/resources';
 
 /**
  * This class represents a basic tracer.
@@ -38,8 +37,10 @@ export class Tracer implements api.Tracer {
   private readonly _generalLimits: GeneralLimits;
   private readonly _spanLimits: SpanLimits;
   private readonly _idGenerator: IdGenerator;
-  readonly resource: IResource;
   readonly instrumentationLibrary: InstrumentationLibrary;
+
+  private readonly _resource: IResource;
+  private readonly _spanProcessor: SpanProcessor;
 
   /**
    * Constructs a new Tracer instance.
@@ -47,14 +48,16 @@ export class Tracer implements api.Tracer {
   constructor(
     instrumentationLibrary: InstrumentationLibrary,
     config: TracerConfig,
-    private _tracerProvider: BasicTracerProvider
+    resource: IResource,
+    spanProcessor: SpanProcessor
   ) {
     const localConfig = mergeConfig(config);
     this._sampler = localConfig.sampler;
     this._generalLimits = localConfig.generalLimits;
     this._spanLimits = localConfig.spanLimits;
     this._idGenerator = config.idGenerator || new RandomIdGenerator();
-    this.resource = _tracerProvider.resource;
+    this._resource = resource;
+    this._spanProcessor = spanProcessor;
     this.instrumentationLibrary = instrumentationLibrary;
   }
 
@@ -138,18 +141,20 @@ export class Tracer implements api.Tracer {
       Object.assign(attributes, samplingResult.attributes)
     );
 
-    const span = new Span(
-      this,
+    const span = new SpanImpl({
+      resource: this._resource,
+      scope: this.instrumentationLibrary,
       context,
-      name,
       spanContext,
-      spanKind,
-      parentSpanId,
+      name,
+      kind: spanKind,
       links,
-      options.startTime,
-      undefined,
-      initAttributes
-    );
+      parentSpanId,
+      attributes: initAttributes,
+      startTime: options.startTime,
+      spanProcessor: this._spanProcessor,
+      spanLimits: this._spanLimits,
+    });
     return span;
   }
 
@@ -249,9 +254,5 @@ export class Tracer implements api.Tracer {
   /** Returns the active {@link SpanLimits}. */
   getSpanLimits(): SpanLimits {
     return this._spanLimits;
-  }
-
-  getActiveSpanProcessor(): SpanProcessor {
-    return this._tracerProvider.getActiveSpanProcessor();
   }
 }